Remove code branches for sklearn versions older than 0.24
