+++ daal4py/sklearn/ensemble/_forest.py
@@ -650,8 +650,7 @@ class RandomForestClassifier(RandomForestClassifier_original, RandomForestBase):
             return super().predict_proba(X)
         X = check_array(X, dtype=[np.float64, np.float32])
         check_is_fitted(self)
-        if sklearn_check_version("0.23"):
-            self._check_n_features(X, reset=False)
+        self._check_n_features(X, reset=False)
         return self._daal_predict_proba(X)
 
     if sklearn_check_version("1.0"):
@@ -670,10 +669,7 @@ class RandomForestClassifier(RandomForestClassifier_original, RandomForestBase):
             if self._cached_estimators_:
                 return self._cached_estimators_
 
-        if sklearn_check_version("0.22"):
-            check_is_fitted(self)
-        else:
-            check_is_fitted(self, "daal_model_")
+        check_is_fitted(self)
         classes_ = self.classes_[0]
         n_classes_ = self.n_classes_[0]
         # convert model to estimators
@@ -1243,10 +1239,7 @@ class RandomForestRegressor(RandomForestRegressor_original, RandomForestBase):
         if hasattr(self, "_cached_estimators_"):
             if self._cached_estimators_:
                 return self._cached_estimators_
-        if sklearn_check_version("0.22"):
-            check_is_fitted(self)
-        else:
-            check_is_fitted(self, "daal_model_")
+        check_is_fitted(self)
         # convert model to estimators
         params = {
             "criterion": self.criterion,
