+++ daal4py/sklearn/metrics/_pairwise.py
@@ -19,6 +19,7 @@ from functools import partial
 
 import numpy as np
 from sklearn.exceptions import DataConversionWarning
+from sklearn.metrics import pairwise_distances as pairwise_distances_original
 from sklearn.metrics.pairwise import (
     _VALID_METRICS,
     PAIRWISE_BOOLEAN_FUNCTIONS,
@@ -69,104 +70,6 @@ def _daal4py_correlation_distance_dense(X):
 def pairwise_distances(
     X, Y=None, metric="euclidean", *, n_jobs=None, force_all_finite=True, **kwds
 ):
-    """ Compute the distance matrix from a vector array X and optional Y.
-
-    This method takes either a vector array or a distance matrix, and returns
-    a distance matrix. If the input is a vector array, the distances are
-    computed. If the input is a distances matrix, it is returned instead.
-
-    This method provides a safe way to take a distance matrix as input, while
-    preserving compatibility with many other algorithms that take a vector
-    array.
-
-    If Y is given (default is None), then the returned matrix is the pairwise
-    distance between the arrays from both X and Y.
-
-    Valid values for metric are:
-
-    - From scikit-learn: ['cityblock', 'cosine', 'euclidean', 'l1', 'l2',
-      'manhattan']. These metrics support sparse matrix
-      inputs.
-      ['nan_euclidean'] but it does not yet support sparse matrices.
-
-    - From scipy.spatial.distance: ['braycurtis', 'canberra', 'chebyshev',
-      'correlation', 'dice', 'hamming', 'jaccard', 'kulsinski', 'mahalanobis',
-      'minkowski', 'rogerstanimoto', 'russellrao', 'seuclidean',
-      'sokalmichener', 'sokalsneath', 'sqeuclidean', 'yule']
-      See the documentation for scipy.spatial.distance for details on these
-      metrics. These metrics do not support sparse matrix inputs.
-
-    Note that in the case of 'cityblock', 'cosine' and 'euclidean' (which are
-    valid scipy.spatial.distance metrics), the scikit-learn implementation
-    will be used, which is faster and has support for sparse matrices (except
-    for 'cityblock'). For a verbose description of the metrics from
-    scikit-learn, see the __doc__ of the sklearn.pairwise.distance_metrics
-    function.
-
-    Read more in the :ref:`User Guide <metrics>`.
-
-    Parameters
-    ----------
-    X : array [n_samples_a, n_samples_a] if metric == "precomputed", or, \
-             [n_samples_a, n_features] otherwise
-        Array of pairwise distances between samples, or a feature array.
-
-    Y : array [n_samples_b, n_features], optional
-        An optional second feature array. Only allowed if
-        metric != "precomputed".
-
-    metric : string, or callable
-        The metric to use when calculating distance between instances in a
-        feature array. If metric is a string, it must be one of the options
-        allowed by scipy.spatial.distance.pdist for its metric parameter, or
-        a metric listed in pairwise.PAIRWISE_DISTANCE_FUNCTIONS.
-        If metric is "precomputed", X is assumed to be a distance matrix.
-        Alternatively, if metric is a callable function, it is called on each
-        pair of instances (rows) and the resulting value recorded. The callable
-        should take two arrays from X as input and return a value indicating
-        the distance between them.
-
-    n_jobs : int or None, optional (default=None)
-        The number of jobs to use for the computation. This works by breaking
-        down the pairwise matrix into n_jobs even slices and computing them in
-        parallel.
-
-        ``None`` means 1 unless in a :obj:`joblib.parallel_backend` context.
-        ``-1`` means using all processors. See :term:`Glossary <n_jobs>`
-        for more details.
-
-    force_all_finite : boolean or 'allow-nan', (default=True)
-        Whether to raise an error on np.inf and np.nan in array. The
-        possibilities are:
-
-        - True: Force all values of array to be finite.
-        - False: accept both np.inf and np.nan in array.
-        - 'allow-nan': accept only np.nan values in array. Values cannot
-          be infinite.
-
-        .. versionadded:: 0.22
-
-    **kwds : optional keyword parameters
-        Any further parameters are passed directly to the distance function.
-        If using a scipy.spatial.distance metric, the parameters are still
-        metric dependent. See the scipy docs for usage examples.
-
-    Returns
-    -------
-    D : array [n_samples_a, n_samples_a] or [n_samples_a, n_samples_b]
-        A distance matrix D such that D_{i, j} is the distance between the
-        ith and jth vectors of the given matrix X, if Y is None.
-        If Y is not None, then D_{i, j} is the distance between the ith array
-        from X and the jth array from Y.
-
-    See also
-    --------
-    pairwise_distances_chunked : performs the same calculation as this
-        function, but returns a generator of chunks of the distance matrix, in
-        order to limit memory usage.
-    paired_distances : Computes the distances between corresponding
-                       elements of two arrays
-    """
     if metric not in _VALID_METRICS and not callable(metric) and metric != "precomputed":
         raise ValueError(
             "Unknown metric %s. Valid metrics are %s, or 'precomputed', "
@@ -250,3 +153,5 @@ if sklearn_check_version("1.3"):
         },
         prefer_skip_nested_validation=True,
     )(pairwise_distances)
+
+pairwise_distances.__doc__ = pairwise_distances_original.__doc__
