+++ sklearnex/neighbors/knn_unsupervised.py
@@ -14,12 +14,6 @@
 # limitations under the License.
 # ===============================================================================
 
-try:
-    from packaging.version import Version
-except ImportError:
-    from distutils.version import LooseVersion as Version
-
-from sklearn import __version__ as sklearn_version
 from sklearn.neighbors._unsupervised import NearestNeighbors as sklearn_NearestNeighbors
 from sklearn.utils.validation import _deprecate_positional_args, check_is_fitted
 
@@ -30,69 +24,12 @@ from onedal.neighbors import NearestNeighbors as onedal_NearestNeighbors
 from .._device_offload import dispatch, wrap_output_data
 from .common import KNeighborsDispatchingBase
 
-if sklearn_check_version("0.22") and Version(sklearn_version) < Version("0.23"):
-
-    class NearestNeighbors_(sklearn_NearestNeighbors):
-        def __init__(
-            self,
-            n_neighbors=5,
-            radius=1.0,
-            algorithm="auto",
-            leaf_size=30,
-            metric="minkowski",
-            p=2,
-            metric_params=None,
-            n_jobs=None,
-        ):
-            super().__init__(
-                n_neighbors=n_neighbors,
-                radius=radius,
-                algorithm=algorithm,
-                leaf_size=leaf_size,
-                metric=metric,
-                p=p,
-                metric_params=metric_params,
-                n_jobs=n_jobs,
-            )
-
-else:
-
-    class NearestNeighbors_(sklearn_NearestNeighbors):
-        if sklearn_check_version("1.2"):
-            _parameter_constraints: dict = {
-                **sklearn_NearestNeighbors._parameter_constraints
-            }
-
-        @_deprecate_positional_args
-        def __init__(
-            self,
-            *,
-            n_neighbors=5,
-            radius=1.0,
-            algorithm="auto",
-            leaf_size=30,
-            metric="minkowski",
-            p=2,
-            metric_params=None,
-            n_jobs=None,
-        ):
-            super().__init__(
-                n_neighbors=n_neighbors,
-                radius=radius,
-                algorithm=algorithm,
-                leaf_size=leaf_size,
-                metric=metric,
-                p=p,
-                metric_params=metric_params,
-                n_jobs=n_jobs,
-            )
-
 
 @control_n_jobs(decorated_methods=["fit", "kneighbors"])
-class NearestNeighbors(NearestNeighbors_, KNeighborsDispatchingBase):
+class NearestNeighbors(sklearn_NearestNeighbors, KNeighborsDispatchingBase):
     __doc__ = sklearn_NearestNeighbors.__doc__
     if sklearn_check_version("1.2"):
-        _parameter_constraints: dict = {**NearestNeighbors_._parameter_constraints}
+        _parameter_constraints: dict = {**sklearn_NearestNeighbors._parameter_constraints}
 
     @_deprecate_positional_args
     def __init__(
@@ -159,18 +96,10 @@ class NearestNeighbors(NearestNeighbors_, KNeighborsDispatchingBase):
             or getattr(self, "_tree", 0) is None
             and self._fit_method == "kd_tree"
         ):
-            if sklearn_check_version("0.24"):
-                sklearn_NearestNeighbors.fit(self, self._fit_X, getattr(self, "_y", None))
-            else:
-                sklearn_NearestNeighbors.fit(self, self._fit_X)
-        if sklearn_check_version("0.22"):
-            result = sklearn_NearestNeighbors.radius_neighbors(
-                self, X, radius, return_distance, sort_results
-            )
-        else:
-            result = sklearn_NearestNeighbors.radius_neighbors(
-                self, X, radius, return_distance
-            )
+            sklearn_NearestNeighbors.fit(self, self._fit_X, getattr(self, "_y", None))
+        result = sklearn_NearestNeighbors.radius_neighbors(
+            self, X, radius, return_distance, sort_results
+        )
 
         return result
